Development

+++ composeApp/src/commonMain/kotlin/org/softwareanvil/domain/models/Syllable.kt
@@ -8,5 +8,5 @@ data class Syllable(
     val type: SyllableType,
     val category: SyllableCategory,
     val value: String,
-    val weight: Int
+    val weight: Long
 )
